Enforce float64 dtype for all GP surrogates and acquisition functions
- Add dtype validation to SingleTaskGPSurrogate and MultiTaskGPSurrogate
  fit() and predict() methods
- Add dtype validation to EI and UCB acquisition forward() methods
- Add _validate_dtype helper to MultiObjectiveAcquisition base class
- Update NEHVI.build() to validate X_baseline and Y dtypes
- Enforce float64 output in Project.get_training_data()
- Add comprehensive dtype validation tests

+++ src/folio/core/project.py
@@ -354,6 +354,7 @@ class Project:
         - The observation is marked as failed (obs.failed is True)
         - The target value cannot be computed (returns None)
         """
+
         target = self.get_target()
         input_names = [inp.name for inp in self.inputs]
         X_rows = []
@@ -367,4 +368,4 @@ class Project:
             row = [obs.inputs[name] for name in input_names]
             X_rows.append(row)
             y_values.append(y)
-        return np.array(X_rows), np.array(y_values)
+        return np.array(X_rows, dtype=np.float64), np.array(y_values, dtype=np.float64)
